divided the module utils into separate modules: render, common, event
