+++ src/utils/event.js
@@ -1,41 +1,5 @@
 import moment from "moment";
-import {insertPosition} from "./const.js";
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case insertPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case insertPosition.BEFOREEND:
-      container.append(element);
-      break;
-    case insertPosition.AFTEREND:
-      container.after(element);
-      break;
-  }
-};
-
-export const renderTemplate = (container, template, place) => {
-  container.insertAdjacentHTML(place, template);
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-export const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-export const getRandomElement = (arr) => arr[(Math.random() * arr.length) | 0];
-
-export const shuffleArray = (arr) => arr.sort(() => 0.5 - Math.random());
+import {getRandomInteger} from "../utils/common.js";
 
 export const generateDate = (maxShiftEventMinutes, startDate = null) => {
   const millisecondsShift = 1000 * 60 * maxShiftEventMinutes;
